♻️refactor: 코드 포매팅 실행, views 구조화 변경
